+++ src/main/java/com/esprit/alphadev/TunisieCamp/repository/ProhibitedWordRepository.java
@@ -3,5 +3,5 @@ package com.esprit.alphadev.TunisieCamp.repository;
 import com.esprit.alphadev.TunisieCamp.entities.ProhibitedWord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProhibitedWordRepository extends JpaRepository<ProhibitedWord, Long> {
+public interface ProhibitedWordRepository extends JpaRepository<ProhibitedWord, Integer> {
 }
