+++ src/main/java/com/esprit/alphadev/TunisieCamp/service/CampingCenterService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface CampingCenterService {
     List<CampingCenter> getAllCampingCenters();
 
-    CampingCenter getCampingCenterById(Long id);
+    CampingCenter getCampingCenterById(Integer id);
